+++ tap_service_titan/streams.py
@@ -12,7 +12,7 @@ from singer_sdk import typing as th  # JSON Schema typing helpers
 from singer_sdk.helpers import types
 from singer_sdk.helpers.jsonpath import extract_jsonpath
 from singer_sdk.pagination import BaseAPIPaginator
-from singer_sdk.streams.core import REPLICATION_FULL_TABLE
+from singer_sdk.streams.core import REPLICATION_FULL_TABLE, REPLICATION_INCREMENTAL
 
 from tap_service_titan.client import (
     ServiceTitanExportStream,
@@ -1898,121 +1898,3 @@ class CapacitiesStream(ServiceTitanStream):
     def path(self) -> str:
         """Return the API path for the stream."""
         return f"/dispatch/v2/tenant/{self._tap.config['tenant_id']}/capacity"
-
-
-class CustomReports(ServiceTitanStream):
-    """Define reviews stream."""
-
-    def __init__(self, *args, **kwargs) -> None:
-        self._report = kwargs.pop("report")
-        super().__init__(
-            *args,
-            **kwargs,
-            name=f"custom_report_{self._report['report_name']}",
-        )
-
-    name = "custom_report"
-    rest_method = "POST"
-    replication_method = REPLICATION_FULL_TABLE
-
-    @staticmethod
-    def _get_datatype(string_type: str) -> th.JSONTypeHelper:
-        mapping = {
-            # String , Number , Boolean , Date , Time
-            "String": th.StringType(),
-            "Number": th.NumberType(),
-            "Boolean": th.BooleanType(),
-            "Date": th.DateTimeType(),
-            "Time": th.StringType(),
-        }
-        return mapping.get(string_type, th.StringType())
-
-    def _get_report_metadata(self) -> dict:
-        report_category = self._report["report_category"]
-        report_id = self._report["report_id"]
-        resp = requests.get(
-            f"{self.url_base}/reporting/v2/tenant/{self.config['tenant_id']}/report-category/{report_category}/reports/{report_id}?pageSize=5000&page=1",
-            headers={**self.http_headers, **self.authenticator.auth_headers},
-            timeout=self.timeout,
-        )
-        resp.raise_for_status()
-        return resp.json()
-
-    @cached_property
-    def schema(self) -> dict:
-        """Get schema.
-
-        Returns:
-            JSON Schema dictionary for this stream.
-        """
-        metadata = self._get_report_metadata()
-        msg = f"Available parameters for custom report `{self._report['report_name']}`: {metadata['parameters']}"
-        self.logger.info(msg)
-        properties: list[th.Property] = [
-            th.Property(field["name"], self._get_datatype(field["dataType"]))
-            for field in metadata["fields"]
-        ]
-
-        return th.PropertiesList(*properties).to_dict()
-
-    @cached_property
-    def path(self) -> str:
-        """Return the API path for the stream."""
-        report_category = self._report["report_category"]
-        report_id = self._report["report_id"]
-        return (
-            f"/reporting/v2/tenant/{self._tap.config['tenant_id']}/report-category"
-            f"/{report_category}/reports/{report_id}/data"
-        )
-
-    def get_url_params(
-        self,
-        context: dict | None,
-        next_page_token: t.Any | None,  # noqa: ANN401
-    ) -> dict[str, t.Any]:
-        """Return a dictionary of values to be used in URL parameterization.
-
-        Args:
-            context: The stream context.
-            next_page_token: The next page index or value.
-
-        Returns:
-            A dictionary of URL query parameters.
-        """
-        params = super().get_url_params(context, next_page_token)
-        params.pop("modifiedOnOrAfter", "")
-        return params
-
-    def prepare_request_payload(
-        self,
-        context: types.Context | None,
-        next_page_token: _TToken | None,
-    ) -> dict | None:
-        """Prepare the data payload for the REST API request.
-
-        By default, no payload will be sent (return None).
-
-        Developers may override this method if the API requires a custom payload along
-        with the request. (This is generally not required for APIs which use the
-        HTTP 'GET' method.)
-
-        Args:
-            context: Stream partition or context dictionary.
-            next_page_token: Token, page number or any request argument to request the
-                next page of data.
-        """
-        return {"parameters": self._report["parameters"]}
-
-    def parse_response(self, response: requests.Response) -> t.Iterable[dict]:
-        """Parse the response and return an iterator of result records.
-
-        Args:
-            response: The HTTP ``requests.Response`` object.
-
-        Yields:
-            Each record from the source.
-        """
-        resp = response.json()
-        field_names = [field["name"] for field in resp["fields"]]
-        for record in resp["data"]:
-            yield dict(zip(field_names, record))
